Update Prayer Request Like Model Types
