--- components/prayer-request/prayer-request-types.ts
@@ -1,11 +1,0 @@
-export type PrayerRequestActionCreateRequest = {
-  userId: number;
-  submittedDate: string;
-};
-
-export type PrayerRequestLikeModel = {
-  prayerRequestLikeId?: number;
-  prayerRequestId?: number;
-  submittedUserId?: number;
-  submittedDate?: string;
-};
